Compile with warnings. Mainly required marking unused parameter, removed unused variables, marking static functions.

+++ src/plasma/memory/cpool.cpp
@@ -49,7 +49,7 @@ void *CPool::alloc(Index cache_index)
 {
     void *buffer = nullptr;
     // check if there is a buffer available in the cache
-    DEBUG_ASSERT((cache_index == INVALID_INDEX) || (cache_index < _n_caches), "invalid cache index " << cache_index);
+    DEBUG_ASSERT((cache_index == INVALID_INDEX) || ((uint32_t)cache_index < _n_caches), "invalid cache index " << cache_index);
     if (cache_index != INVALID_INDEX && _cache_heads[cache_index] != INVALID_INDEX) {
         DEBUG_ASSERT_OP(_cache_counts[cache_index], >, 0, "free list isn't empty though count equals 0");
         buffer = index_to_address(_cache_heads[cache_index]);
@@ -75,7 +75,7 @@ void CPool::free_address(Index cache_index, void *buffer)
 
 void CPool::free(Index cache_index, Index buffer_index)
 {
-    DEBUG_ASSERT((cache_index == INVALID_INDEX) || (cache_index < _n_caches), "invalid cache index " << cache_index);
+    DEBUG_ASSERT((cache_index == INVALID_INDEX) || ((uint32_t)cache_index < _n_caches), "invalid cache index " << cache_index);
     if (cache_index != INVALID_INDEX && _cache_counts[cache_index] < _max_buffers_per_cache) {
         // return the buffer to the cache
         void *buffer = index_to_address(buffer_index);
